Refactored to have better architecture
